requirements.txt
fastapi==0.103.0
uvicorn==0.22.0
transformers==4.33.0
Pillow==9.4.0
opencv-python==4.8.0.76
numpy==1.26.0
pyyaml==6.0
torch==2.1.0
sentencepiece
protobuf
python-multipart